refactor configuration view
