+++ FullScreenTranslator/Localizable.xcstrings
@@ -61,7 +61,14 @@
       }
     },
     "Settings" : {
-
+      "localizations" : {
+        "ja" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "設定"
+          }
+        }
+      }
     },
     "Silent duration" : {
       "localizations" : {
